requirements.txt
fastai==2.0.16
fastcore==1.2.2
https://download.pytorch.org/whl/cpu/torch-1.6.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.7.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
streamlit==0.72.0
numpy==1.19.0
pillow~=7.2
python-multipart==0.0.5
timm==0.2.1
wwf==0.0.4
